move mock into gamestate. It's only used by this package so move it to the relevant package.
