+++ cardinal/ecs/gamestate/mock.go
@@ -1,4 +1,4 @@
-package storage
+package gamestate
 
 import (
 	"fmt"
